refactor: move category types and validation logic to shared package
- Move CategoryGroup and Category types to packages/shared/src/types/categories.ts
- Move isActivityFullyCategorized() logic to packages/shared/src/logic/browsable.ts
- Update all imports across app and api workspaces to use the shared package
- Export new types and logic from packages/shared/src/index.ts

This allows category-related code to be reused across the platform without duplication.
