+++ apps/api/src/types.ts
@@ -103,19 +103,6 @@ export type UserInfoWithEmail = UserInfo & {
   isEditor?: boolean;
 };
 
-export type CategoryGroup = {
-  name: string;
-  isRequired: boolean;
-  isExclusive: boolean;
-  categories: Category[];
-};
-
-export type Category = {
-  code: string;
-  description: string;
-  term: string;
-};
-
 export type ContentClassification = {
   id: number;
   code: string;
