+++ apps/app/src/paths/editor/EditorHeader.tsx
@@ -44,13 +44,7 @@ import { IoGitBranch } from "react-icons/io5";
 import { LuCircleHelp, LuLibraryBig } from "react-icons/lu";
 
 import axios from "axios";
-import {
-  AssignmentStatus,
-  ContentType,
-  ContentDescription,
-  CategoryGroup,
-  Category,
-} from "../../types";
+import { AssignmentStatus, ContentType, ContentDescription } from "../../types";
 import { contentTypeToName, getIconInfo } from "../../utils/activity";
 import { SiteContext } from "../SiteHeader";
 import { getDiscourseUrl } from "../../utils/discourse";
@@ -63,12 +57,16 @@ import { LibraryEditorControls } from "../../widgets/editor/LibraryEditorControl
 import { editorUrl } from "../../utils/url";
 import { NameBar } from "../../widgets/NameBar";
 import { loader as settingsLoader } from "./EditorSettingsMode";
-import { isActivityFullyCategorized } from "../../utils/classification";
 import { useIframeMenuDismissOverlay } from "../../utils/useIframeMenuDismissOverlay";
 import { MenuDismissOverlay } from "../../components/MenuDismissOverlay";
 import { IFRAME_MENU_IDS } from "../../utils/iframeMenuIds";
 import { useControlledMenu } from "../../utils/useControlledMenu";
 import { useMenuTooltipSuppression } from "../../utils/useMenuTooltipSuppression";
+import {
+  Category,
+  CategoryGroup,
+  isActivityFullyCategorized,
+} from "@doenet-tools/shared";
 
 export async function loader({
   params,
