+++ packages/shared/src/index.ts
@@ -1,4 +1,6 @@
 export * from "./types/activityViewer.js";
+export * from "./types/categories.js";
 export * from "./utils/ipfs.js";
 export * from "./utils/test.js";
 export * from "./apiTypes.js";
+export * from "./logic/browsable.js";
