+++ packages/shared/src/types/categories.ts
@@ -1,0 +1,12 @@
+export type CategoryGroup = {
+  name: string;
+  isRequired: boolean;
+  isExclusive: boolean;
+  categories: Category[];
+};
+
+export type Category = {
+  code: string;
+  description: string;
+  term: string;
+};
